+++ BE/src/main/java/com/codesquad/airbnb/controller/WishListController.java
@@ -1,6 +1,6 @@
 package com.codesquad.airbnb.controller;
 
-import com.codesquad.airbnb.dto.PropertiesResponseDTO;
+import com.codesquad.airbnb.dto.property.PropertiesResponseDTO;
 import com.codesquad.airbnb.service.PropertyService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
