+++ BE/src/main/java/com/codesquad/airbnb/dao/PropertyDAO.java
@@ -5,8 +5,8 @@ import com.codesquad.airbnb.dao.rowMapper.PropertyDTORowMapper;
 import com.codesquad.airbnb.dao.rowMapper.PropertyRowMapper;
 import com.codesquad.airbnb.domain.Property;
 import com.codesquad.airbnb.dto.HostDTO;
-import com.codesquad.airbnb.dto.PropertyDTO;
-import com.codesquad.airbnb.dto.PropertyDetailResponseDTO;
+import com.codesquad.airbnb.dto.property.PropertyDTO;
+import com.codesquad.airbnb.dto.property.PropertyDetailResponseDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
