+++ BE/src/main/java/com/codesquad/airbnb/dto/reservation/ReservationRequestDTO.java
@@ -1,4 +1,4 @@
-package com.codesquad.airbnb.dto;
+package com.codesquad.airbnb.dto.reservation;
 
 import java.time.LocalDate;
